+++ src/Application.Core/Features/Article/Queries/GetArticle/GetArticleQueryResponse.cs
@@ -16,7 +16,7 @@ namespace Application.Core.Features.Article.Queries.GetArticle
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public string Image { get; set; } = string.Empty;
-        public double BasePrice { get; set; }
+        public decimal BasePrice { get; set; }
         public Guid CategoryId { get; set; }
         public string CategoryName { get; set; } = string.Empty;
         public bool Publish { get; set; }
